requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-extensions==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.1
reportlab==3.5.48
sqlparse==0.3.1
Werkzeug==1.0.1
zipp==3.1.0
django-heroku